requirements.txt
python-dotenv 
schedule
PyGithub
requests